+++ connect/runtime/src/main/java/org/apache/kafka/connect/converters/LongConverter.java
@@ -14,30 +14,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.connect.storage;
+package org.apache.kafka.connect.converters;
 
+import org.apache.kafka.common.serialization.LongDeserializer;
 import org.apache.kafka.common.serialization.LongSerializer;
-import org.apache.kafka.common.serialization.Serializer;
 import org.apache.kafka.connect.data.Schema;
+import org.apache.kafka.connect.storage.Converter;
+import org.apache.kafka.connect.storage.HeaderConverter;
 
-public class LongConverterTest extends NumberConverterTest<Long> {
+/**
+ * {@link Converter} and {@link HeaderConverter} implementation that only supports serializing to and deserializing from long values.
+ * It does support handling nulls. When converting from bytes to Kafka Connect format, the converter will always return an
+ * optional INT64 schema.
+ * <p>
+ * This implementation currently does nothing with the topic names or header names.
+ */
+public class LongConverter extends NumberConverter<Long> {
 
-    public Long[] samples() {
-        return new Long[]{Long.MIN_VALUE, 1234L, Long.MAX_VALUE};
+    public LongConverter() {
+        super("long", Schema.OPTIONAL_INT64_SCHEMA, new LongSerializer(), new LongDeserializer());
     }
 
-    @Override
-    protected Schema schema() {
-        return Schema.OPTIONAL_INT64_SCHEMA;
-    }
-
-    @Override
-    protected NumberConverter<Long> createConverter() {
-        return new LongConverter();
-    }
-
-    @Override
-    protected Serializer<Long> createSerializer() {
-        return new LongSerializer();
-    }
 }
