+++ connect/runtime/src/main/java/org/apache/kafka/connect/runtime/isolation/PluginUtils.java
@@ -127,11 +127,6 @@ public class PluginUtils {
             + "|json\\..*"
             + "|file\\..*"
             + "|converters\\..*"
-            + "|storage\\.DoubleConverter"
-            + "|storage\\.FloatConverter"
-            + "|storage\\.IntegerConverter"
-            + "|storage\\.LongConverter"
-            + "|storage\\.ShortConverter"
             + "|storage\\.StringConverter"
             + "|rest\\.basic\\.auth\\.extension\\.BasicAuthSecurityRestExtension"
             + ")"
